feat: rename settings options
