+++ .env.example
@@ -4,7 +4,7 @@ EMAIL_HOST_USER="notifications@gmail.com"
 EMAIL_HOST_PASSWORD="password"
 # Django configurations
 DJANGO_SECRET_KEY="some-secret"
-DJANGO_SETTINGS_MODULE="bd_api.settings.dev"
+DJANGO_SETTINGS_MODULE="bd_api.settings.local"
 # Database
 DB_HOST="localhost"
 DB_PORT="5432"
