+++ .github/workflows/cd-dev.yaml
@@ -102,7 +102,7 @@ jobs:
             envFrom:
               - secretRef:
                   name: api-development-secrets
-            settingsModule: "bd_api.settings.prod"
+            settingsModule: "bd_api.settings.remote"
           database:
             host: "cloud-sql-proxy"
             port: 5432
