+++ .github/workflows/cd-staging.yaml
@@ -76,7 +76,7 @@ jobs:
             envFrom:
               - secretRef:
                   name: api-staging-secrets
-            settingsModule: "bd_api.settings.prod"
+            settingsModule: "bd_api.settings.remote"
           database:
             host: "cloud-sql-proxy"
             port: 5432
