+++ bd_api/settings/__init__.py
@@ -1,2 +1,2 @@
 # -*- coding: utf-8 -*-
-from .dev import *  # noqa
+from .local import *  # noqa
